[DX] Extract translation models in the new component
