+++ spec/Pim/Bundle/TransformBundle/Normalizer/Flat/TranslationNormalizerSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\TransformBundle\Normalizer\Flat;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Entity\AttributeTranslation;
-use Pim\Bundle\TranslationBundle\Entity\TranslatableInterface;
+use Akeneo\Component\Localization\Model\TranslatableInterface;
 
 class TranslationNormalizerSpec extends ObjectBehavior
 {
