+++ spec/Pim/Bundle/TransformBundle/Normalizer/Structured/TranslationNormalizerSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\TransformBundle\Normalizer\Structured;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Entity\AttributeTranslation;
-use Pim\Bundle\TranslationBundle\Entity\TranslatableInterface;
+use Akeneo\Component\Localization\Model\TranslatableInterface;
 
 class TranslationNormalizerSpec extends ObjectBehavior
 {
