+++ src/Pim/Bundle/CatalogBundle/Entity/Category.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\CatalogBundle\Entity;
 use Akeneo\Component\Classification\Model\Category as BaseCategory;
 use Doctrine\Common\Collections\ArrayCollection;
 use JMS\Serializer\Annotation\ExclusionPolicy;
-use Pim\Bundle\TranslationBundle\Entity\AbstractTranslation;
+use Akeneo\Component\Localization\Model\AbstractTranslation;
 use Pim\Component\Catalog\Model\CategoryInterface;
 
 /**
