+++ src/Pim/Bundle/CatalogBundle/Entity/GroupType.php
@@ -4,7 +4,7 @@ namespace Pim\Bundle\CatalogBundle\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use JMS\Serializer\Annotation\ExclusionPolicy;
-use Pim\Bundle\TranslationBundle\Entity\AbstractTranslation;
+use Akeneo\Component\Localization\Model\AbstractTranslation;
 use Pim\Component\Catalog\Model\GroupTypeInterface;
 
 /**
