+++ src/Pim/Bundle/TranslationBundle/EventListener/AddLocaleListener.php
@@ -4,7 +4,7 @@ namespace Pim\Bundle\TranslationBundle\EventListener;
 
 use Doctrine\Common\EventSubscriber;
 use Doctrine\ORM\Event\LifecycleEventArgs;
-use Pim\Bundle\TranslationBundle\Entity\TranslatableInterface;
+use Akeneo\Component\Localization\Model\TranslatableInterface;
 
 /**
  * Aims to inject user context locale into translatable entities
