+++ src/Pim/Bundle/VersioningBundle/UpdateGuesser/TranslationsUpdateGuesser.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\VersioningBundle\UpdateGuesser;
 use Akeneo\Component\Versioning\Model\VersionableInterface;
 use Doctrine\Common\Util\ClassUtils;
 use Doctrine\ORM\EntityManager;
-use Pim\Bundle\TranslationBundle\Entity\AbstractTranslation;
+use Akeneo\Component\Localization\Model\AbstractTranslation;
 
 /**
  * Translation update guesser
