+++ src/Pim/Component/Catalog/Model/AbstractAttribute.php
@@ -5,7 +5,7 @@ namespace Pim\Component\Catalog\Model;
 use Doctrine\Common\Collections\ArrayCollection;
 use Pim\Bundle\CatalogBundle\AttributeType\AbstractAttributeType;
 use Pim\Bundle\CatalogBundle\AttributeType\AttributeTypes;
-use Pim\Bundle\TranslationBundle\Entity\AbstractTranslation;
+use Akeneo\Component\Localization\Model\AbstractTranslation;
 use Pim\Component\Catalog\Model\LocaleInterface;
 
 /**
