+++ src/Pim/Component/Catalog/Model/AttributeGroupInterface.php
@@ -4,7 +4,7 @@ namespace Pim\Component\Catalog\Model;
 
 use Akeneo\Component\Versioning\Model\VersionableInterface;
 use Doctrine\Common\Collections\ArrayCollection;
-use Pim\Bundle\TranslationBundle\Entity\TranslatableInterface;
+use Akeneo\Component\Localization\Model\TranslatableInterface;
 
 /**
  * Attribute Group interface
