+++ src/Pim/Component/Catalog/Model/AttributeInterface.php
@@ -4,7 +4,7 @@ namespace Pim\Component\Catalog\Model;
 
 use Akeneo\Component\Versioning\Model\VersionableInterface;
 use Doctrine\Common\Collections\ArrayCollection;
-use Pim\Bundle\TranslationBundle\Entity\TranslatableInterface;
+use Akeneo\Component\Localization\Model\TranslatableInterface;
 use Pim\Component\Catalog\Model\LocaleInterface;
 use Symfony\Component\Validator\GroupSequenceProviderInterface;
