+++ src/Pim/Component/Catalog/Model/CategoryInterface.php
@@ -4,7 +4,7 @@ namespace Pim\Component\Catalog\Model;
 
 use Akeneo\Component\Classification\Model\CategoryInterface as BaseCategoryInterface;
 use Akeneo\Component\Versioning\Model\VersionableInterface;
-use Pim\Bundle\TranslationBundle\Entity\TranslatableInterface;
+use Akeneo\Component\Localization\Model\TranslatableInterface;
 
 /**
  * Category interface
